Move layout-engine.js and graph.js to folder "common"
